pyproject.toml
[project]
name = "testcontainers-demo"
version = "0.1.0"
description = "Fibonacci queue worker demonstrating testcontainers with Python"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "nats-py>=2.7.0",
    "redis>=5.0.0",
    "psycopg[binary]>=3.1.0",
    "boto3>=1.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "testcontainers[postgres,redis,nats,localstack]>=4.7.0",
]
